requirements.txt
tensorflow==2.4.1
opencv-python==4.5.1.48
numpy==1.19.5
pandas==1.2.3
eventlet==0.30.2
Flask==1.1.2
seaborn==0.11.1
matplotlib==3.3.4
tensorboard==2.4.1
pywin32==300
psutil==5.8.0
Pillow==8.1.2
python-socketio==4.6.1
python-engineio==3.13.2
